add: vdom with function h()
now the component returns vdom element, and is used to build real dom node during render()
