+++ index.html
@@ -7,7 +7,7 @@
     <title>Like React</title>
   </head>
   <body>
-    <div id="app"></div>
+    <div id="root"></div>
     <script type="module" src="/src/main.js"></script>
   </body>
 </html>
