+++ src/core/createDomElement.js
@@ -1,0 +1,22 @@
+export default function createDomElement(vnode) {
+    const { type, props, children } = vnode;
+
+    const el = document.createElement(type);
+
+    for (const [key, value] of Object.entries(props || {})) {
+        if (key.startsWith('on') && typeof value === 'function') {
+            const event = key.slice(2).toLowerCase();
+            el.addEventListener(event, value);
+        }
+    }
+
+    children.flat().forEach(child => {
+        if (typeof child === 'string' || typeof child === 'number') {
+            el.appendChild(document.createTextNode(child));
+        } else if (child instanceof Node) {
+            el.appendChild(child);
+        }
+    });
+
+    return el;
+}
