--- src/core/createElement.js
@@ -1,20 +1,0 @@
-export function createElement(type, props = {}, ...children) {
-  const el = document.createElement(type);
-
-  for (const [key, value] of Object.entries(props || {})) {
-    if (key.startsWith('on') && typeof value === 'function') {
-      const event = key.slice(2).toLowerCase();
-      el.addEventListener(event, value);
-    }
-  }
-
-  children.flat().forEach(child => {
-    if (typeof child === 'string' || typeof child === 'number') {
-      el.appendChild(document.createTextNode(child));
-    } else if (child instanceof Node) {
-      el.appendChild(child);
-    }
-  });
-
-  return el;
-}
