+++ src/core/h.js
@@ -1,0 +1,7 @@
+export function h(type, props = {}, ...children) {
+  return {
+    type,
+    props,
+    children: children.flat()
+  };
+}
