+++ src/main.js
@@ -2,6 +2,6 @@ import './style.css'
 import { render } from './core/render';
 import { App } from './App';
 
-const root = document.getElementById("app");
+const root = document.getElementById("root");
 console.log('in main.js')
 render(App, root);
